Created separate iOS and Mac versions of NavigationTitleExampleView
